(from AES) dgroup: Fix a race where sometimes things don't get logged
@esmet noticed a CI flake where `Example_timeout` failed because it was
missing the "shutting down (not-so-gracefully)..." log line.  If all
workers quit *very quickly* after being given the shutdown signal, there's
a race where the logger might not log that the signal was ever sent.

I'd run the tests with `go test -count=1000` to try to catch things like
that, but some how it didn't catch that.

Since this is a pretty subtle race, I honestly have no idea how to add a
test for it that would more reliably catch it than the current test does.