dexec: Improve log output
The dexec log format predates dlog.  It packs everything in to a line of
human-friendly but unstructured data.  Change things around to be nicer
with the structured logging we get with dlog.

Additionally (to make the above acceptable) fuss with the default fallback
dlog logger to print the dexec fields in a better order.