dcontext: Add a WithoutCancel function for non-timed-out cleanup
The idea+implementation is borrowed from Telepresence.  The tests are new.

A long time ago I had a conversation with Rafael Schloming about "what
about non-timed-out cleanup?" in the context of dgroup.

Well, much later, it turned out that Telepresence's answer to that was
withoutCancel{}.  So let's put that answer into dlib.