Add a LoggerWithMaxLevel interface
This interface is meant to be used by `Logger` implementations that
will discard messages above a certain log level. A prominent example
is the current logrus wrapper.

The interface is motivated by the need to sometimes skip resource
consuming string formatting for messages that the logger will discard.
Example:

    if dlog.MaxLogLevel(ctx) >= dlog.LogLevelTrace {
      // use a string buffer to build a trace message from some network
      // package, then log it with
      dlog.Trace(ctx, msg)
    }

The max-level can also be usedful when an alternative logger must
be created for some third party module that implements its own logger
API.